64b port
git-svn-id: svn+ssh://jhoberock-dt/home/jhoberock/svnbase/jared/src/gotham@1142 3bb10773-34f4-0310-b88c-ec9d511ba0d9

+++ renderers/VarianceRenderer.h
@@ -59,9 +59,9 @@ class VarianceRenderer
      *  \param n The number of samples generated.
      *  \param samples The non-zero sample values.
      */
-    virtual void depositVariance(const unsigned int px,
-                                 const unsigned int py,
-                                 const unsigned int n,
+    virtual void depositVariance(const size_t px,
+                                 const size_t py,
+                                 const size_t n,
                                  const std::vector<Spectrum> &samples);
 
     virtual void depositVariance(const PathSampler::HyperPoint &x,
